# ⚡️ Speed up method `Usage.opentelemetry_attributes` by 85%